+++ AIProject/iCo/Core/Remote/HTTP/Endpoint.swift
@@ -84,7 +84,4 @@ extension Requestable {
     }
 }
 
-enum HTTPMethod: String {
-    case get = "GET"
-    case post = "POST"
-}
+
