+++ AIProject/iCo/Data/API/FearGreed/FearGreedAPIService.swift
@@ -7,19 +7,6 @@
 
 import Foundation
 
-enum FearGreedEndpoint {
-    case main
-}
-
-extension FearGreedEndpoint: Requestable {
-    var baseURL: String { "https://api.alternative.me" }
-    var path: String { "/fng" }
-    var httpMethod: HTTPMethod { .get }
-    var queryParameters: Encodable? { nil }
-    var bodyParameters: Encodable? { nil }
-    var headers: [String : String] { [:] }
-}
-
 /// 공포 탐욕 지수 API 관련 서비스를 제공합니다.
 final class FearGreedAPIService: FearGreedProvider {
     let network: NetworkClient
